+++ torchmetrics/regression/mean_squared_error.py
@@ -39,6 +39,8 @@ class MeanSquaredError(Metric):
             before returning the value at the step. default: False
         process_group:
             Specify the process group on which synchronization is called. default: None (which selects the entire world)
+        squared:
+            If True returns MSE value, if False returns RMSE value.
 
     Example:
         >>> from torchmetrics import MeanSquaredError
@@ -56,6 +58,7 @@ class MeanSquaredError(Metric):
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
         dist_sync_fn: Callable = None,
+        squared: bool = True,
     ):
         super().__init__(
             compute_on_step=compute_on_step,
@@ -66,6 +69,7 @@ class MeanSquaredError(Metric):
 
         self.add_state("sum_squared_error", default=tensor(0.0), dist_reduce_fx="sum")
         self.add_state("total", default=tensor(0), dist_reduce_fx="sum")
+        self.squared = squared
 
     def update(self, preds: Tensor, target: Tensor):
         """
@@ -84,7 +88,7 @@ class MeanSquaredError(Metric):
         """
         Computes mean squared error over state.
         """
-        return _mean_squared_error_compute(self.sum_squared_error, self.total)
+        return _mean_squared_error_compute(self.sum_squared_error, self.total, squared=self.squared)
 
     @property
     def is_differentiable(self):